テストエラー対策の修正（APIパスの変更など）
