--- app/controllers/pets_controller.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-# Swagger Petstore
-#
-# No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
-#
-# The version of the OpenAPI document: 1.0.0
-# Generated by: https://github.com/openapitools/openapi-generator.git
-#
-class PetsController < ApplicationController
-  def index
-    # Your code here
-
-    # render json: {"message" => "yes, it worked"}
-    # render json: [{ 'id' => 1, 'name' => 'コノベル太郎' }]
-    pets = Sentence.all
-    render json: pets
-  end
-end
