+++ app/controllers/v1/sentences_controller.rb
@@ -1,0 +1,40 @@
+# frozen_string_literal: true
+
+# Sentences API
+#
+# No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
+#
+# The version of the OpenAPI document: 1.0.0
+# Generated by: https://github.com/openapitools/openapi-generator.git
+#
+
+module V1
+  # SentencesController
+  class SentencesController < ApplicationController
+    def show
+      result = FindSentenceService.new(params[:sentence_id]).call
+      if result[:success]
+        render_success_response(result[:sentence])
+      else
+        render_error_response(result[:error])
+      end
+    end
+
+    private
+
+    def render_success_response(sentence)
+      render json: {
+        main: {
+          sentence_id: sentence.id,
+          sentence: sentence.sentence,
+          created_at: sentence.created_at,
+          updated_at: sentence.updated_at
+        }
+      }, status: :ok
+    end
+
+    def render_error_response(error)
+      render json: { error: }, status: :not_found
+    end
+  end
+end
