+++ spec/support/request_spec_helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module RequestSpecHelper
+  def json
+    JSON.parse(response.body)
+  end
+end
